+++ sythil_saas_server/__openerp__.py
@@ -1,6 +1,6 @@
 {
     'name': "Sythil SAAS Server",
-    'version': "1.0.4",
+    'version': "1.0.5",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",